+++ NEWS.md
@@ -1,6 +1,14 @@
-# dv.tables 0.0.3.9002
+
+# dv.tables 0.1.0
 
 * Add Tplyr_table module 
+* Add jumping functionality
+* Implement label truncation
+* Implement functionality to download to Word (.rtf) format
+* Allow subsetting of hierarchy and group variables drop-down list through module argument
+* Remove subject identifier variable from hierarchy and group variables drop-down list
+* Alert when group selection also selected in hierarchy
+* Allow a maximum of 2 hierarchy variable selections
 
 # dv.tables 0.0.3